Changed how expressions and loop closures are wrapped.
They're now substituted into a block, as opposed to an expression.  This
means you can actually use *statements* without having to do a block
yourself.

Also, closures are now defined and constrained at the top of `main`.  This
means you can define variables *outside* the closure, which is kinda nice.