Removed content hashing for file inputs.
The *original* idea was to hash contents so that a script, even if moved,
could be cached.  What is *far more* common, at least thus far, is making
changes to a particular script and re-running it in one place.  Since
content hashing made this more painful than it should be, it got removed.

Also, the way the digest for file paths was created was changed from
DEFLATE to SHA-1.  The problem was that, upon actually *testing* the idea,
it turns out that `hex(deflate(path))` *always* makes the result longer.
As such, it only really captures the *start* of the path.

SHA-1, then, has a *much* better chance of actually giving scripts in the
same directory different digests.

Finally, the digest limit was dropped to 16 nibbles; that works out to
64 bits, which *should* be enough for our purposes.  Even if we *do* get
a collision, the worst thing that will happen is an unwanted recompilation.