Changed how input splitting is done.
All the manifest-extraction stuff is now in its own module.  Also, the
arm that used to do it for files has been moved out to a function.  This
is being done in preparation for more manifest formats.

The way hashbangs were stripped was also changed.  It's now a function,
which is just *way* easier than that line iterator business.  It's also
an excuse to introduce the multi-pattern pattern.