Replaced Pattern code with regexen.
This was the last remaining unstable feature being used by
`cargo-script`, so it should now build on 1.3 stable.

The only known problem is that due to the way the cargo shipped with
rustc <= 1.3 compiles build scripts, we won't build under multirust-rs.
I do not intend to fix this; it's easier to just wait for 1.4 to be
released.