Fix `PathExt::is_file` not being called.
Turns out that this method hasn't been called *at all* since 1.5.  It
was being shadowed by the "native" `is_file` that was added, but the
compiler never mentioned this.  1.10 improves the unused code lint to
correctly identify this.

Rather than go to a full version detection build script setup for this
one method, I've just changed the name to something more unique.