Mark tests which need internet access
Tests which depend on downloading dependency crates will fail when executed in
an environment which has no internet access. To allow for offline testing,
these tests are only enabled with the "online_tests" feature.

To run all tests, including those which need internet access, it is now
necessary to run:

    cargo test --features online_tests